backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
pydantic==2.8.2
SQLAlchemy==2.0.32
aiosqlite==0.20.0
rapidfuzz==3.8.1
httpx==0.27.0
azure-ai-documentintelligence==1.0.0b3
azure-identity==1.17.1
azure-storage-blob==12.22.0
python-dotenv==1.0.1
openai==1.43.0
gunicorn==21.2.0
